+++ examples/runtime.rs
@@ -20,7 +20,7 @@ fn main() {
             .connect_to_node("redis://127.0.0.1:6379")
             .desired_pool_size(10)
             .reservation_limit(None) // No limit
-            .checkout_mode(Immediately) // No timeout
+            .default_checkout_mode(Immediately) // No timeout
             //.task_executor(runtime.executor()) no explicit executor!
             .finish_redis_rs()
             .unwrap();